Move gateway HTTP proxy tests into package structure
