+++ tracdap-services/tracdap-gateway/src/test/java/org/finos/tracdap/gateway/proxy/http/Http1Client.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Accenture Global Solutions Limited
+ * Copyright 2023 Accenture Global Solutions Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.finos.tracdap.gateway;
+package org.finos.tracdap.gateway.proxy.http;
 
 import io.netty.bootstrap.Bootstrap;
 import io.netty.buffer.Unpooled;
